fixed PR feedback
