+++ src/script/components/CellsGlobalView/CellsTable/CellsTableColumns/CellsTableRowOptions/CellsTableRowOptions.tsx
@@ -20,10 +20,8 @@
 import {DropdownMenu, MoreIcon} from '@wireapp/react-ui-kit';
 
 import {openFolder} from 'Components/CellsGlobalView/common/openFolder/openFolder';
-import {useFileHistoryModal} from 'Components/Modals/FileHistoryModal/hooks/useFileHistoryModal';
 import {CellsRepository} from 'Repositories/cells/CellsRepository';
 import {CellNode, CellNodeType} from 'src/script/types/cellNode';
-import {isFileEditable} from 'Util/FileTypeUtil';
 import {t} from 'Util/LocalizerUtil';
 import {forcedDownloadFile} from 'Util/util';
 
@@ -34,17 +32,14 @@ import {showShareModal} from '../CellsShareModal/CellsShareModal';
 
 interface CellsTableRowOptionsProps {
   node: CellNode;
-
   cellsRepository: CellsRepository;
 }
 
 export const CellsTableRowOptions = ({node, cellsRepository}: CellsTableRowOptionsProps) => {
   const {handleOpenFile} = useCellsFilePreviewModal();
-  const {showModal} = useFileHistoryModal();
 
   const url = node.url;
   const name = node.type === CellNodeType.FOLDER ? `${node.name}.zip` : node.name;
-  const isEditable = node.type === CellNodeType.FILE && isFileEditable(node.extension!);
 
   return (
     <DropdownMenu>
@@ -63,14 +58,6 @@ export const CellsTableRowOptions = ({node, cellsRepository}: CellsTableRowOptio
         <DropdownMenu.Item onClick={() => showShareModal({type: node.type, uuid: node.id, cellsRepository})}>
           {t('cells.options.share')}
         </DropdownMenu.Item>
-        {isEditable && (
-          <>
-            <DropdownMenu.Item onClick={() => handleOpenFile(node, true)}>{t('cells.options.edit')}</DropdownMenu.Item>
-            <DropdownMenu.Item onClick={() => showModal(node.id, () => handleOpenFile(node, false))}>
-              {t('cells.options.versionHistory')}
-            </DropdownMenu.Item>
-          </>
-        )}
         {!!url && (
           <DropdownMenu.Item
             onClick={() =>
